double

+++ utils/src/main/java/com/epam/utils/StringUtils.java
@@ -3,6 +3,6 @@ package com.epam.utils;
 public class StringUtils {
     public static boolean isPositiveNumber(String str) {
         //code!
-        return Integer.parseInt(str) > 0;
+        return Double.parseDouble(str) > 0;
     }
 }
